+++ src/main/robot/Roues.java
@@ -9,11 +9,11 @@ public class Roues extends Robot {
         this.baseVitesse = vitesse;
     }
 
-    public int getVitesse() {
+    public double getVitesse() {
         return getVitesse(this.position);
     }
 
-    public int getVitesse(Case position) {
+    public double getVitesse(Case position) {
         return this.baseVitesse;
     }
